dagster_pixar/pyproject.toml
[project]
name = "dagster_pixar"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9,<3.14"

[project.optional-dependencies]
dev = [
    "dagster-webserver",
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.dagster]
module_name = "dagster_pixar.definitions"
code_location_name = "dagster_pixar"